fix(0286): resolve Tier-2 dual-role imports — extract _tradition_style and _crossref_qa
test_no_tier3_entry_point_is_imported flagged the two new script-to-script
imports. Per the 0250/0254 pattern: tradition labels/colors move to the
neutral scripts/_tradition_style.py (plot_fig_traditions re-exports them
unchanged); wilson_ci + fetch_crossref_refs move to scripts/qa/_crossref_qa.py
(qa_citations imports them; qa_metadata keeps its own wilson_ci untouched).
Make/dvc deps updated. Validation: citer-limited figure re-renders
byte-identical; canonical figure old-vs-new pixel diff (0.09% px) is within
the pre-existing run-to-run nondeterminism envelope (0.02% px on repeat
renders of the SAME script) — anti-aliasing jitter, no content change.
