+++ scripts/figures/plot_fig_traditions.py
@@ -33,25 +33,13 @@ from utils import get_logger
 
 log = get_logger("plot_fig_traditions")
 
-TRADITION_LABELS = {
-    "pricing": "Environmental economics\n(pricing & quantities)",
-    "cdm":     "Development economics\n(CDM & carbon markets)",
-    "unfccc":  "Burden-sharing\n(UNFCCC & institutions)",
-    "other":   None,
-}
-
-TRADITION_COLORS = {
-    "pricing": "#1a6496",
-    "cdm":     "#e07b39",
-    "unfccc":  "#4a9e6b",
-    "other":   "#DDDDDD",
-}
-TRADITION_EDGE_COLORS = {
-    "pricing": "#1a6496",
-    "cdm":     "#e07b39",
-    "unfccc":  "#4a9e6b",
-    "other":   "#CCCCCC",
-}
+# Shared with the citer-limited variant via the neutral style module
+# (Tier-2 surface rule, ticket 0286); names re-exported here unchanged.
+from _tradition_style import (
+    TRADITION_COLORS,
+    TRADITION_EDGE_COLORS,
+    TRADITION_LABELS,
+)
 
 
 def _render_traditions(G, partition, pos, comm_to_tradition,
